safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFEConvolveMatrixElement-svgdom-preserveAlpha-prop-expected.txt

CONSOLE MESSAGE: line 41: TypeError: 'undefined' is not an object (evaluating 'convolveMatrixElement.preserveAlpha.baseVal = true')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'preserveAlpha' property of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL convolveMatrixElement.preserveAlpha.baseVal should be false. Threw exception TypeError: 'undefined' is not an object (evaluating 'convolveMatrixElement.preserveAlpha.baseVal')
